+++ nalgebra-lapack/CHANGELOG.md
@@ -12,6 +12,7 @@ For the **changes between versions 0.4.0 and 0.27.0** see the main
 * add column-pivoting QR decomposition and solver
 * fix logic error in calculation of complex eigenvalues in eigen-decomposition.
 * change the feature flags for choosing the lapack backend, update docs accordingly
+* removed untested complex support in QR decomposition
 
 ## [0.4.0] - 2016-09-07